announcement command added
