+++ telegram_bot_service/src/main/resources/application.yml
@@ -8,12 +8,6 @@
   admin:
     allowed_users:
       Marandyuk_Anatolii,
-      HomerTheBald,
-      shapran_dmitry,
-      Galactic_Alex,
-      coderror404,
-      Shainara,
-      Kulpinov_Evgeny,
       Denis_Shvartcer
 
 #настройка веб адреса приложения
